+++ mediatailor/src/main/java/com/bitmovin/player/integration/mediatailor/AdPlaybackEventEmitter.kt
@@ -29,41 +29,51 @@ internal class DefaultAdPlaybackEventEmitter(
                 when {
                     previousPlayingAdBreak == null && playingAdBreak != null -> {
                         eventEmitter.emit(MediaTailorEvent.AdBreakStarted(playingAdBreak.adBreak))
-                        eventEmitter.emit(
-                            MediaTailorEvent.AdStarted(
-                                ad = playingAdBreak.ad,
-                                indexInQueue = playingAdBreak.adIndex,
-                            ),
-                        )
+                        playingAdBreak.ad?.let { ad ->
+                            eventEmitter.emit(
+                                MediaTailorEvent.AdStarted(
+                                    ad = ad,
+                                    indexInQueue = playingAdBreak.adIndex,
+                                ),
+                            )
+                        }
                     }
 
                     previousPlayingAdBreak != null && playingAdBreak != null &&
                         previousPlayingAdBreak.adBreak.id != playingAdBreak.adBreak.id -> {
-                        eventEmitter.emit(MediaTailorEvent.AdFinished(previousPlayingAdBreak.ad))
+                        previousPlayingAdBreak.ad?.let { previousAd ->
+                            eventEmitter.emit(MediaTailorEvent.AdFinished(previousAd))
+                        }
                         eventEmitter.emit(MediaTailorEvent.AdBreakFinished(previousPlayingAdBreak.adBreak))
                         eventEmitter.emit(MediaTailorEvent.AdBreakStarted(playingAdBreak.adBreak))
-                        eventEmitter.emit(
-                            MediaTailorEvent.AdStarted(
-                                ad = playingAdBreak.ad,
-                                indexInQueue = playingAdBreak.adIndex,
-                            ),
-                        )
+                        playingAdBreak.ad?.let { ad ->
+                            eventEmitter.emit(
+                                MediaTailorEvent.AdStarted(
+                                    ad = ad,
+                                    indexInQueue = playingAdBreak.adIndex,
+                                ),
+                            )
+                        }
                     }
 
-                    playingAdBreak != null && previousPlayingAdBreak?.ad?.id != playingAdBreak.ad.id -> {
-                        if (previousPlayingAdBreak?.ad != null) {
-                            eventEmitter.emit(MediaTailorEvent.AdFinished(previousPlayingAdBreak.ad))
+                    playingAdBreak != null && previousPlayingAdBreak?.ad?.id != playingAdBreak.ad?.id -> {
+                        previousPlayingAdBreak?.ad?.let { previousAd ->
+                            eventEmitter.emit(MediaTailorEvent.AdFinished(previousAd))
                         }
-                        eventEmitter.emit(
-                            MediaTailorEvent.AdStarted(
-                                ad = playingAdBreak.ad,
-                                indexInQueue = playingAdBreak.adIndex,
-                            ),
-                        )
+                        playingAdBreak.ad?.let { ad ->
+                            eventEmitter.emit(
+                                MediaTailorEvent.AdStarted(
+                                    ad = ad,
+                                    indexInQueue = playingAdBreak.adIndex,
+                                ),
+                            )
+                        }
                     }
 
                     previousPlayingAdBreak != null && playingAdBreak == null -> {
-                        eventEmitter.emit(MediaTailorEvent.AdFinished(previousPlayingAdBreak.ad))
+                        previousPlayingAdBreak.ad?.let { ad ->
+                            eventEmitter.emit(MediaTailorEvent.AdFinished(ad))
+                        }
                         eventEmitter.emit(MediaTailorEvent.AdBreakFinished(previousPlayingAdBreak.adBreak))
                     }
                 }